docs(PPDSC-2063): created internal accordion component
